pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "openai>=1.99.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "telegram>=0.0.1",
]
